requirements.txt
matplotlib==3.3.3
tensorflow==2.4.0
numpy==1.19.5
opencv_python==4.5.1.48
Pillow==8.0.1
PyQt5==5.15.2
